centralise fixtures into conftest.py; move generic subclasses of ABCs to utils/generic_subclasses.py
