+++ cstar/tests/unit_tests/base/test_input_dataset.py
@@ -6,106 +6,7 @@ from unittest import mock
 
 import pytest
 
-from cstar.base import InputDataset
-from cstar.base.datasource import DataSource
-
-
-class MockInputDataset(InputDataset):
-    """Mock subclass of the InputDataset abstract base class.
-
-    Since InputDataset is an abstract base class, this mock class is needed to allow
-    instantiation for testing purposes. It inherits from InputDataset without adding any
-    new behavior, serving only to allow tests to create and manipulate instances.
-    """
-
-    pass
-
-
-@pytest.fixture
-def local_input_dataset():
-    """Fixture to provide a mock local InputDataset instance.
-
-    This fixture patches properties of the DataSource class to simulate a local dataset,
-    initializing it with relevant attributes like location, start date, and end date.
-
-    Mocks
-    -----
-    - Mocked DataSource.location_type property returning 'path'
-    - Mocked DataSource.source_type property returning 'netcdf'
-    - Mocked DataSource.basename property returning 'local_file.nc'
-
-    Yields
-    ------
-    MockInputDataset: Instance representing a local input dataset for testing.
-    """
-    with (
-        mock.patch.object(
-            DataSource, "location_type", new_callable=mock.PropertyMock
-        ) as mock_location_type,
-        mock.patch.object(
-            DataSource, "source_type", new_callable=mock.PropertyMock
-        ) as mock_source_type,
-        mock.patch.object(
-            DataSource, "basename", new_callable=mock.PropertyMock
-        ) as mock_basename,
-    ):
-        mock_location_type.return_value = "path"
-        mock_source_type.return_value = "netcdf"
-        mock_basename.return_value = "local_file.nc"
-
-        dataset = MockInputDataset(
-            location="some/local/source/path/local_file.nc",
-            start_date="2024-10-22 12:34:56",
-            end_date="2024-12-31 23:59:59",
-        )
-
-        yield dataset
-
-
-@pytest.fixture
-def remote_input_dataset():
-    """Fixture to provide a mock remote InputDataset instance.
-
-    This fixture patches properties of the DataSource class to simulate a remote dataset,
-    initializing it with attributes such as URL location, file hash, and date range.
-
-    Mocks
-    -----
-    - Mocked DataSource.location_type property returning 'url'
-    - Mocked DataSource.source_type property returning 'netcdf'
-    - Mocked DataSource.basename property returning 'remote_file.nc'
-
-    Yields
-    ------
-    MockInputDataset: Instance representing a remote input dataset for testing.
-    """
-    # Using context managers to patch properties on DataSource
-    with (
-        mock.patch.object(
-            DataSource, "location_type", new_callable=mock.PropertyMock
-        ) as mock_location_type,
-        mock.patch.object(
-            DataSource, "source_type", new_callable=mock.PropertyMock
-        ) as mock_source_type,
-        mock.patch.object(
-            DataSource, "basename", new_callable=mock.PropertyMock
-        ) as mock_basename,
-    ):
-        # Mock property return values for a remote file (URL)
-        mock_location_type.return_value = "url"
-        mock_source_type.return_value = "netcdf"
-        mock_basename.return_value = "remote_file.nc"
-
-        # Create the InputDataset instance; it will use the mocked DataSource
-        dataset = MockInputDataset(
-            location="http://example.com/remote_file.nc",
-            file_hash="abc123",
-            start_date="2024-10-22 12:34:56",
-            end_date="2024-12-31 23:59:59",
-        )
-
-        # Yield the dataset for use in the test
-        yield dataset
+from cstar.tests.unit_tests.generic_abc_subclasses import MockInputDataset
 
 
 class TestInputDatasetInit:
