+++ cstar/tests/unit_tests/roms/test_roms_external_codebase.py
@@ -9,16 +9,6 @@ from cstar.roms.external_codebase import ROMSExternalCodeBase
 from cstar.system.manager import cstar_sysmgr
 
 
-@pytest.fixture
-def roms_codebase():
-    """Fixture providing a configured instance of `ROMSExternalCodeBase` for testing."""
-    source_repo = "https://github.com/CWorthy-ocean/ucla-roms.git"
-    checkout_target = "246c11fa537145ba5868f2256dfb4964aeb09a25"
-    return ROMSExternalCodeBase(
-        source_repo=source_repo, checkout_target=checkout_target
-    )
-
-
 def test_default_source_repo(roms_codebase):
     """Test if the default source repo is set correctly."""
     assert (
